Use 32-bit floats for all real-time audio calculations
* Avoid using doubles in general (i.e., Complex number implementation)
* Update unit tests

+++ src/Jukebox/Chunk.hpp
@@ -43,6 +43,6 @@ namespace Dynamo {
          * @brief Frame offset of the chunk
          *
          */
-        double frame;
+        float frame;
     };
 } // namespace Dynamo
